Keep generator storage classes in separate file
