+++ scripts/py_matter_idl/examples/matter_idl_plugin/__init__.py
@@ -14,7 +14,8 @@
 
 import os
 
-from matter.idl.generators import CodeGenerator, GeneratorStorage
+from matter.idl.generators import CodeGenerator
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.matter_idl_types import Cluster, Command, Field, Idl
