+++ scripts/py_matter_idl/matter/idl/data_model_xml_parser.py
@@ -26,8 +26,8 @@ except ImportError:
     sys.path.append(str(Path(__file__).resolve().parent / ".." / ".."))
     from matter.idl.data_model_xml import ParseSource, ParseXmls
 
-from matter.idl.generators import GeneratorStorage
 from matter.idl.generators.idl import IdlGenerator
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.matter_idl_parser import CreateParser
 from matter.idl.matter_idl_types import Idl
