+++ scripts/py_matter_idl/matter/idl/generators/cpp/application/application_generator.py
@@ -16,8 +16,9 @@ import os
 from dataclasses import dataclass
 from typing import List, Mapping, Optional
 
-from matter.idl.generators import CodeGenerator, GeneratorStorage
+from matter.idl.generators import CodeGenerator
 from matter.idl.generators.cluster_selection import server_side_clusters
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.generators.type_definitions import TypeLookupContext
 from matter.idl.matter_idl_types import Bitmap, Idl, ServerClusterInstantiation
