+++ scripts/py_matter_idl/matter/idl/generators/cpp/sdk/sdk_generator.py
@@ -14,7 +14,8 @@
 #
 import os
 
-from matter.idl.generators import CodeGenerator, GeneratorStorage
+from matter.idl.generators import CodeGenerator
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.matter_idl_types import (AccessPrivilege, Attribute, AttributeQuality, Command, CommandQuality, FieldQuality, Idl,
                                          Struct)
