+++ scripts/py_matter_idl/matter/idl/generators/cpp/tlvmeta/__init__.py
@@ -16,7 +16,8 @@ import os
 from dataclasses import dataclass
 from typing import Generator, List, Optional
 
-from matter.idl.generators import CodeGenerator, GeneratorStorage
+from matter.idl.generators import CodeGenerator
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.matter_idl_types import Cluster, Field, Idl, StructTag
