+++ scripts/py_matter_idl/matter/idl/test_idl_generator.py
@@ -27,8 +27,8 @@ except ImportError:
     sys.path.append(str(Path(__file__).resolve().parent / ".." / ".."))
     from matter.idl.matter_idl_parser import CreateParser
 
-from matter.idl.generators import GeneratorStorage
 from matter.idl.generators.idl import IdlGenerator
+from matter.idl.generators.storage import GeneratorStorage
 from matter.idl.matter_idl_types import Idl
