+++ scripts/py_matter_idl/matter/idl/zapxml_parser.py
@@ -29,8 +29,8 @@ except ImportError:
 if __name__ == '__main__':
     import click
 
-    from matter.idl.generators import GeneratorStorage
     from matter.idl.generators.idl import IdlGenerator
+    from matter.idl.generators.storage import GeneratorStorage
 
     class InMemoryStorage(GeneratorStorage):
         def __init__(self):
